Move Hecke-related things to ext
