+++ src/ecc/codes/classical/lifted.jl
@@ -1,82 +1,7 @@
-import Nemo: characteristic, lift, matrix_repr
-import Hecke: representation_matrix, GroupAlgebra, GroupAlgebraElem, GroupElem, abelian_group, group_algebra
-
-
-"""
-Classical codes lifted over a group algebra [panteleev2021degenerate](@cite) [panteleev2022asymptotically](@cite).
-
-- `A::Matrix`: the base matrix of the code, whose elements are in a group algebra.
-- `repr::Function`: a function that converts a group algebra element to a matrix; default to be [`permutation_repr`](@ref) for GF(2)-algebra.
-
-TODO why we need such a freedom of representation?
-
-The parity-check matrix is constructed by applying `repr` to each element of `A`, which is mathematically a linear map from a group algebra element to a binary matrix.
-This will enlarge the parity check matrix from `A` with each element being inflated into a matrix. The procedure is called a lift [panteleev2022asymptotically](@cite).
-
-See also: [`LPCode`](@ref).
-"""
-struct LiftedCode <: ClassicalCode
-    A::GroupAlgebraElemMatrix
-    GA::GroupAlgebra
-    repr::Function
-
-    function LiftedCode(A::GroupAlgebraElemMatrix; GA::GroupAlgebra=parent(A[1,1]), repr::Function)
-        all(elem.parent == GA for elem in A) || error("The base ring of all elements in the code must be the same as the group algebra")
-        new(A, GA, repr)
+function LiftedCode(args...; kwargs...)
+    ext = Base.get_extension(QuantumClifford, :QuantumCliffordHeckeExt)
+    if isnothing(ext)
+        throw("The `LiftedCode` depends on the package `AbstractAlgebra` and `Hecke` but you have not installed or imported them yet. Immediately after you import `AbstractAlgebra` and `Hecke`, the `LiftedCode` will be available.")
     end
+    return ext.LiftedCode(args...; kwargs...)
 end
-
-default_repr(y::GroupAlgebraElem{FqFieldElem, <: GroupAlgebra}) = Matrix((x -> Bool(Int(lift(ZZ, x)))).(representation_matrix(y)))
-
-"""
-The GroupAlgebraElem with `GF(2)` coefficients can be converted to a permutation matrix by `representation_matrix` provided by Hecke.
-"""
-function LiftedCode(A::Matrix{GroupAlgebraElem{FqFieldElem, <: GroupAlgebra}}; GA::GroupAlgebra=parent(A[1,1]))
-    !(characteristic(base_ring(A[1, 1])) == 2) && error("The default permutation representation applies only to GF(2) group algebra; otherwise, a custom representation function should be provided")
-    LiftedCode(A; GA=GA, repr=default_repr)
-end
-
-function LiftedCode(group_elem_array::Matrix{<: GroupOrAdditiveGroupElem}; GA::GroupAlgebra=group_algebra(GF(2), parent(group_elem_array[1,1])), repr::Union{Function, Nothing}=nothing)
-    A = zeros(GA, size(group_elem_array)...)
-    for i in axes(group_elem_array, 1), j in axes(group_elem_array, 2)
-        A[i, j] = GA[A[i, j]]
-    end
-    if repr === nothing
-        return LiftedCode(A; GA=GA, repr=default_repr)
-    else
-        return LiftedCode(A; GA=GA, repr=repr)
-    end
-end
-
-function LiftedCode(shift_array::Matrix{Int}, l::Int; GA::GroupAlgebra=group_algebra(GF(2), abelian_group(l)))
-    A = zeros(GA, size(shift_array)...)
-    for i in 1:size(shift_array, 1)
-        for j in 1:size(shift_array, 2)
-            A[i, j] = GA[shift_array[i, j]%l+1]
-        end
-    end
-    return LiftedCode(A; GA=GA, repr=default_repr)
-end
-
-function lift(repr::Function, mat::GroupAlgebraElemMatrix)
-    vcat([hcat([repr(mat[i, j]) for j in axes(mat, 2)]...) for i in axes(mat, 1)]...)
-end
-
-function parity_checks(c::LiftedCode)
-    h = lift(c.repr, c.A)
-    rk = mod2rank(h) # TODO mod2rank and canonicalize, which is more efficient?
-    rk < size(h, 1) && @warn "The lifted code has redundant rows"
-    return h
-end
-
-code_n(c::LiftedCode) = size(c.A, 2) * size(c.repr(parent(c.A[1,1])(0)), 2)
-
-function mod2rank(h::Matrix{<:Integer}) # TODO move this mod2rank to a common place
-    Z2, _ = residue_ring(ZZ, 2)
-    S = matrix_space(Z2, size(h)...)
-    rank(S(h))
-end
-
-code_s(c::LiftedCode) = mod2rank(parity_checks(c))
-
-code_k(c::LiftedCode) = code_n(c) - code_s(c)
